Upgrade to JOGL2 RC9
JOGL2 is necessary, as JOGL1 is no longer maintained, as necessary for
changes in operating systems, such as the recent changes in Mac OS X
10.8.

The good news is that all of the native code libraries are now found in
jar files, so we no longer need a lib/ directory in the JTK home. Note
that the javarun and jythonrun scripts have been changed to specify the
new names for the JOGL jar files.