Switch to the Apache 2 license
This was accomplished using the following steps:

1) Replaced the license.txt file with the Apache 2 license text.

2) Updated existing license blurbs:

       perl -0777 -i -pe 's/([ \/*<>!-]+)\n\s*Copyright \(c\) (\d\d\d\d.*) All rights reserved.\n.*cpl-v10.html\n([ \/*<>!-]+)\n/\1\nCopyright \2\nLicensed under the Apache License, Version 2.0 \(the "License"\);\nyou may not use this file except in compliance with the License.\nYou may obtain a copy of the License at\n\n    http:\/\/www.apache.org\/licenses\/LICENSE-2.0\n\nUnless required by applicable law or agreed to in writing, software\ndistributed under the License is distributed on an "AS IS" BASIS,\nWITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.\nSee the License for the specific language governing permissions and\nlimitations under the License.\n\3\n/igs' $(git grep -l 'Copyright')

3) Checked for remaining instances of the Common Public License text:

       git grep Common

   There were only two:
   - src/demo/jython/edu/mines/jtk/dsp/SteerablePyramidDemo.py
   - src/main/java/edu/mines/jtk/ogl/MakeGl.java

4) Manually updated the above two files.

I explored using the license-gradle-plugin [1], which would have been
quite nice to not only update the license headers, but also keep them up
to date. However, it is easiest when using such automated tooling to
standardize the headers with same text across all files -- for example,
by using a copyright year range, with complete list of copyright holders
across the entire project, rather than tracking per-file copyright years
and holders. But since that might be a contentious change, I settled on
a simple search-and-replace instead. A motivated developer could
probably bend the license-gradle-plugin into the right shape to preserve
the existing copyright years and authors on a per-file basis.

[1] https://github.com/hierynomus/license-gradle-plugin